move EVRP packages from EV to DVRP (now DVRP depends on EV)
