+++ contribs/dvrp/pom.xml
@@ -33,6 +33,11 @@
 		</dependency>
 		<dependency>
 			<groupId>org.matsim.contrib</groupId>
+			<artifactId>ev</artifactId>
+			<version>14.0-SNAPSHOT</version>
+		</dependency>
+		<dependency>
+			<groupId>org.matsim.contrib</groupId>
 			<artifactId>locationchoice</artifactId>
 			<version>14.0-SNAPSHOT</version>
 		</dependency>
