+++ contribs/ev/pom.xml
@@ -13,7 +13,7 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.matsim.contrib</groupId>
-			<artifactId>dvrp</artifactId>
+			<artifactId>common</artifactId>
 			<version>14.0-SNAPSHOT</version>
 		</dependency>
 		<dependency>
