--- contribs/ev/src/main/java/org/matsim/contrib/ev/dvrp/ChargingTask.java
@@ -1,35 +1,0 @@
-/* *********************************************************************** *
- * project: org.matsim.*
- * *********************************************************************** *
- *                                                                         *
- * copyright       : (C) 2018 by the members listed in the COPYING,        *
- *                   LICENSE and WARRANTY file.                            *
- * email           : info at matsim dot org                                *
- *                                                                         *
- * *********************************************************************** *
- *                                                                         *
- *   This program is free software; you can redistribute it and/or modify  *
- *   it under the terms of the GNU General Public License as published by  *
- *   the Free Software Foundation; either version 2 of the License, or     *
- *   (at your option) any later version.                                   *
- *   See also COPYING, LICENSE and WARRANTY file                           *
- *                                                                         *
- * *********************************************************************** */
-
-package org.matsim.contrib.ev.dvrp;
-
-import org.matsim.contrib.ev.charging.ChargingWithAssignmentLogic;
-import org.matsim.contrib.ev.fleet.ElectricVehicle;
-
-/**
- * @author michalm
- */
-public interface ChargingTask extends ETask {
-	ChargingWithAssignmentLogic getChargingLogic();
-
-	ElectricVehicle getElectricVehicle();
-
-	double getChargingStartedTime();
-
-	void setChargingStartedTime(double chargingStartedTime);
-}
